Change the logic to determine whether buyers can see the inline guest experience based on merchants config and ELMO results
Prior to this change, we didn't handle the case when merchants don't
provide `enableInlineGuest`

This change will this logic

If `enableInlineGuest` is `true`, buyers ALWAYS see the inline guest.
If `enableInlineGuest` is `false`, buyers NEVER see the inline guest.
If `enableInlineGuest` is `undefined` (merchant doens't pass it), we
take the results from ELMO.